refactor(game-map): extract some other common logic
